some improvements

+++ RayTracingGPU/RayTracingGPU/interval_cuda.h
@@ -32,12 +32,12 @@ const static Interval universe(-infinity, +infinity);
 
 inline float random_float() {
     // Returns a random real in [0,1).
-    return  (float) (rand() / (RAND_MAX + 1.0));
+    return  rand() / (RAND_MAX + 1.0);
 }
 
 inline float random_float(float min, float max) {
     // Returns a random real in [min,max).
-    return (float)(min + (max - min) * random_float());
+    return min + (max - min) * random_float();
 }
 
 #endif
